Convert the Appeal class to a model class
This removes the dual nature of the Appeal constructor ("from db" or "not from
db").  Instead, the constructor will either take no arguments at all and
construct an empty object, or take a map from prefixed database column names to
values, and assign those values to the appropriate properties.

The validate() method has been turned into an instance method that validates the
state of the object.

The insert() method is no longer called from the constructor, and it now calls
validate() as its first step.

Any queries that were executed during the constructor have now been moved to
lazy-loading getter methods so that they are executed only when necessary.